# zhiyi-authorization/src/main/resources/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "使用教程: https://zhiyi.zone",
    "version": "1.0",
    "title": "ZhiYi-Cloud"
  },
  "host": "localhost:6628",
  "basePath": "/",
  "tags": [
    {
      "name": "oauth2",
      "description": "OAuth2 Grant"
    }
  ],
  "paths": {
    "/oauth/token": {
      "post": {
        "tags": [
          "oauth2"
        ],
        "summary": "OAuth2 登录",
        "operationId": "oauth2Token",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "grant_type",
            "in": "query",
            "description": "password,refresh_token,authorization_code,social",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "query",
            "description": "grant_type=social",
            "required": false,
            "type": "string"
          },
          {
            "name": "code",
            "in": "query",
            "description": "grant_type=social,authorization_code",
            "required": false,
            "type": "string"
          },
          {
            "name": "username",
            "in": "query",
            "description": "grant_type=password",
            "required": false,
            "type": "string"
          },
          {
            "name": "password",
            "in": "query",
            "description": "grant_type=password",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "BearerToken": [
              "global"
            ]
          }
        ],
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/captcha/image": {
      "get": {
        "tags": [
          "oauth2"
        ],
        "summary": "图形验证码",
        "operationId": "oauth2Captcha",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "随机字符串",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "BearerToken": [
              "global"
            ]
          }
        ],
        "deprecated": false,
        "x-order": "2147483641"
      }
    }
  },
  "securityDefinitions": {
    "BearerToken": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
  }
}